+++ nidmviewerfsl/tests/testError.py
