+++ queries/askIfOboStatistic.txt
