+++ queries/selectExcursionSetDetails.txt
