+++ queries/selectStatisticType.txt
