+++ queries/selectVersionNum.txt
